feat: profile 분리
